Refactor: modularize GUI into separate tab components
