+++ .github/workflows/build-python-app.yml
@@ -10,11 +10,13 @@ on:
       - "action-test"
       - "feature/*"
       - "feat-*"
+      - "dev*"
   pull_request:
     branches:
       - "master"
       - "action-test"
       - "feature/*"
+      - "dev*"
 
 permissions:
   contents: read
@@ -48,7 +50,7 @@ jobs:
 
     - name: Build with PyInstaller
       run: |
-        pyinstaller -F --windowed -i docs/images/fish-cake.ico --name DG-LAB-VRCOSC --add-data "docs/images/fish-cake.ico;docs/images" src/app.py
+        pyinstaller -F --windowed -i docs/images/fish-cake.ico --name DG-LAB-VRCOSC --add-data "docs/images/fish-cake.ico;docs/images" --paths=./src src/app.py
       shell: pwsh
 
     - name: Generate commit hash and timestamp file
